refactor: merged edit functions into one endpoint
